+++ packages/dataviews/src/dataviews/index.tsx
@@ -41,7 +41,7 @@ import DataViewsViewConfig, {
 } from '../components/dataviews-view-config';
 import normalizeFields from '../field-types';
 import useData from '../hooks/use-data';
-import { useInfiniteScrollData } from '../components/dataviews-layouts/utils/useInfiniteScrollData';
+import { useInfiniteScrollData } from '../hooks/use-infinite-scroll-data';
 import type { Action, Field, View, SupportedLayouts } from '../types';
 import type { SelectionOrUpdater } from '../types/private';
 type ItemWithId = { id: string };
