+++ packages/dataviews/src/hooks/index.ts
@@ -2,3 +2,5 @@
  * Internal dependencies
  */
 export { default as useFormValidity } from './use-form-validity';
+export { useInfiniteScrollData } from './use-infinite-scroll-data';
+export { default as useSelectedItems } from './use-selected-items';
